requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
pyserial>=3.5
ultralytics>=8.0.0
